requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py@7857a845989e421d591c7d231b756433342d00a8